+++ src/main/web/subjective/subjective.js
@@ -31,6 +31,7 @@ const subjective_module = {}
     let _currentTeam;
     let _scoreEntryBackPage;
     let _categoryColumnMapping;
+    let _server_checks_running = 0;
 
     function _init_variables() {
         _subjectiveCategories = {};
@@ -809,7 +810,7 @@ const subjective_module = {}
         subjective_module.uploadData = function(scoresSuccess, scoresFail, judgesSuccess, judgesFail,
             loadSuccess, loadFail) {
 
-            subjective_module.checkServerStatus(function() {
+            subjective_module.checkServerStatus(true, function() {
                 fetch("CheckAuth").then(checkJsonResponse).then(function(data) {
                     if (data.authenticated) {
                         subjective_module
@@ -885,30 +886,39 @@ const subjective_module = {}
         /**
          * Check if the server is reachable.
          *
+         * @param force_check if true, then execute the check even if another one is pending
          * @param onlineCallback function to execute if the server is reachable
          * @param offlineCallback function to execute if the server is not reachable 
          */
-        subjective_module.checkServerStatus = function(onlineCallback, offlineCallback) {
+        subjective_module.checkServerStatus = function(force_check, onlineCallback, offlineCallback) {
+            if (!force_check && _server_checks_running > 0) {
+                subjective_module.log("Skipping server check while another one is active and force is not true");
+                return;
+            }
+            ++_server_checks_running;
+
             subjective_module.log("Checking server status");
 
-            const FETCH_TIMEOUT = 1000; // milliseconds            
+            const FETCH_TIMEOUT = 35000; // milliseconds, need 35 seconds for chromebooks (not sure why)            
             const controller = new AbortController();
             const timeout = setTimeout(function() {
                 controller.abort('Request timed out');
             }, FETCH_TIMEOUT);
 
-            const request = new Request("../images/blank.gif", {
+
+            subjective_module.log("Executing fetch");
+            fetch("../images/blank.gif", {
+                method: 'HEAD',
                 headers: new Headers({
                     'pragma': 'no-cache',
                     'cache-control': 'no-cache'
                 }),
                 cache: "no-store",
                 signal: controller.signal
-            });
+            }).
+                then((response) => {
+                    subjective_module.log("Response received");
 
-            subjective_module.log("Executing fetch: " + request);
-            fetch(request)
-                .then(function(response) {
                     // Clear the timeout as cleanup
                     clearTimeout(timeout);
                     if (!response.ok) {
@@ -919,13 +929,16 @@ const subjective_module = {}
                         onlineCallback();
                     }
                     // no need to resolve the response object as we know the server is online, we don't need the data in the response
-                })
-                .catch(function(err) {
+                }).
+                catch((err) => {
                     subjective_module.log('Server offline (fetch error): ' + err);
 
                     // Error: response error, request timeout or runtime error
                     subjective_module.log("server offline: " + err);
                     offlineCallback();
+                }).
+                finally(() => {
+                    _server_checks_running = Math.min(0, --_server_checks_running);
                 });
         },